Switch from int to long, check payload in BackendConnector

+++ shared/src/main/java/com/sap/psr/vulas/shared/util/FileSearch.java
@@ -43,7 +43,7 @@ public class FileSearch extends AbstractFileSearch {
 
 	private String[] suffixes = null;
 	
-	private int maxSize = -1;
+	private long maxSize = -1;
 
 	/**
 	 * <p>Constructor for FileSearch.</p>
@@ -60,7 +60,7 @@ public class FileSearch extends AbstractFileSearch {
 	 * @param _size maximum accepted file size (no limit, if a value <= 0 is provided)
 	 * @throws java.lang.IllegalArgumentException if any.
 	 */
-	public FileSearch(String[] _s, int _size) throws IllegalArgumentException {
+	public FileSearch(String[] _s, long _size) throws IllegalArgumentException {
 		if(_s==null || _s.length==0)
 			throw new IllegalArgumentException("At least one file extension must be provided");
 		this.suffixes = _s.clone();
